requirements.txt
arrow==0.15.2
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Django==2.2.5
django-cors-headers==3.1.1
django-filter==2.2.0
django-google-maps==0.11.0
django-leaflet==0.24.0
django-rest-knox==4.1.0
djangorestframework==3.10.3
Flask==1.1.1
geos==0.2.2
googlemaps==3.1.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
optional-django==0.3.0
Pillow==6.1.0
postgis==1.0.4
proj==0.1.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycparser==2.19
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2019.2
react==4.3.0
requests==2.22.0
six==1.12.0
spatialite==0.0.3
sqlparse==0.3.0
urllib3==1.25.3
Werkzeug==0.16.0
